feat: introduce oracle utilities and constants for trade guard functionality
